fix: changed f32 to f64 xd

+++ include/matgen/core/types.h
@@ -57,7 +57,7 @@ typedef u64 matgen_index_t;
  * Currently using matgen_value_t precision (f64).
  * Change to f32 for single precision, or make it templated.
  */
-typedef f32 matgen_value_t;
+typedef f64 matgen_value_t;
 
 /**
  * @brief Size type for counts (nnz, capacity, etc.)
